# openapi/akine-api.yaml
openapi: 3.1.0
info:
  contact:
    name: AKINE
    url: https://github.com/Chapadiex/appKine-api
  description: |-
    Contrato REST canonico de AKINE: SaaS multi-tenant para centros de kinesiologia, fisioterapia, rehabilitacion y actividades de salud.

    Este artefacto es propiedad del repositorio backend. El frontend genera su cliente TypeScript desde una version fijada de este contrato y no define DTO manuales.
  title: AKINE API
  version: 0.10.0
servers:
- description: Servidor que sirve este contrato
  url: /
...
paths:
...
  /api/v1/organizations:
    post:
      description: |-
        Crea la organizacion, su primer consultorio y su suscripcion en una sola transaccion. Reservado a la administracion de plataforma.

        El primer consultorio se crea SIEMPRE, incluso por este camino: el contexto de trabajo es Organizacion + Consultorio, asi que un tenant sin ninguna sede no ofreceria ningun contexto seleccionable y quien entrara despues no tendria donde trabajar.

        Este NO es el camino del alta self-service. Esa crea ademas la cuenta y la membership del fundador en la misma transaccion, y la resuelve el modulo de identidad en 01.02.

        El header Idempotency-Key es obligatorio: identifica el intento, de modo que un reintento por timeout de red quede correlacionado en el log y en la auditoria en vez de parecer un alta distinta. En 01.01 el alta duplicada la cierra la unicidad del slug con 409 conflict, no un replay con 200: este camino administrativo no registra idempotencia porque no tiene propietario que registrar, y el registro de onboarding exige uno.
      operationId: createOrganization
      parameters:
      - description: "Identificador unico del intento, generado por el cliente. Reintentar\
          \ la misma alta debe reusar el mismo valor."
        example: 0f9d5f6e-1c2b-4a3d-9e8f-7a6b5c4d3e2f
        in: header
        name: Idempotency-Key
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationResponse"
          description: Organizacion creada. La cabecera Location apunta al recurso
            nuevo
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: "Campos invalidos, o falta el header Idempotency-Key"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: La operacion esta reservada a la administracion de plataforma
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: El plan indicado no existe o ya no es contratable
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: "El slug ya pertenece a otra organizacion, o la clave de idempotencia\
            \ se reuso con otro contenido"
      summary: Alta administrativa de una organizacion
      tags:
      - Organizaciones
...
  /api/v1/organizations/{orgId}/colaborador-invitaciones:
    get:
      description: |-
        Requiere colaborador:manage, y no colaborador:read: lo que se ve aca son direcciones de correo de personas que todavia no aceptaron nada. No es la lista de colaboradores, es la lista de a quien se le escribio.

        Sin paginar a proposito: una organizacion tiene decenas de invitaciones en toda su vida, no miles, y el filtro por estado es lo que hace util al listado.

        Una invitacion vencida figura PENDIENTE con vencida=true. Expirar no es una decision de nadie y por eso no se materializa como estado: se deriva del reloj en cada lectura.
      operationId: listColaboradorInvitaciones
      parameters:
      - in: path
        name: orgId
        required: true
        schema:
          type: integer
          format: int64
      - description: Filtra por estado. Ausente devuelve todas
        in: query
        name: estado
        required: false
        schema:
          type: string
          enum:
          - PENDIENTE
          - ACEPTADA
          - RECHAZADA
          - CANCELADA
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvitacionResponse"
          description: "Invitaciones de la organizacion, de la mas nueva a la mas\
            \ vieja"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: "Sin sesion, sin contexto activo o sin colaborador:manage"
      summary: Listar las invitaciones de la organizacion
      tags:
      - Invitaciones
    post:
      description: |-
        Requiere colaborador:manage sobre el alcance pedido. La organizacion sale del contexto de trabajo activo.

        A diferencia del alta directa, la persona invitada PUEDE NO TENER CUENTA: la crea al aceptar, en el mismo acto, sin un segundo correo de activacion. El vinculo nace cuando acepta, no ahora.

        Emitir no consume cupo del plan; aceptar si. La contracara declarada es que cinco invitaciones pendientes con un solo lugar libre significan que cuatro reciben 409 al aceptar: entra el que llega primero.

        Solo puede haber UNA invitacion pendiente por persona y alcance. Si el enlace vencio o no llego, el camino es reenviar, no emitir otra.
      operationId: createColaboradorInvitacion
      parameters:
      - in: path
        name: orgId
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitacionRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitacionResponse"
          description: Invitacion emitida y correo encolado
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: "Falta el email o el rol, o el rol no es un valor legal"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: "Sin sesion, sin contexto de trabajo activo (missing-tenant-context),\
            \ o sin colaborador:manage sobre el alcance pedido. Nunca 401"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: La sede pedida no es de la organizacion del contexto
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: "Ya hay una invitacion pendiente igual (invitacion-pendiente-duplicada),\
            \ o la persona ya trabaja en la organizacion (colaborador-ya-vinculado)"
      summary: Invitar a alguien a colaborar
      tags:
      - Invitaciones
  /api/v1/organizations/{orgId}/colaborador-invitaciones/{invitacionId}/cancel:
    post:
      description: |-
        El motivo es obligatorio: cancelar es una decision del administrador sobre alguien a quien ya le escribio, y tiene que responder por que seis meses despues. El rechazo del invitado, en cambio, no exige motivo.

        La cancelacion es terminal y libera el email: se puede volver a invitar a esa persona mas adelante.
      operationId: cancelColaboradorInvitacion
      parameters:
      - in: path
        name: orgId
        required: true
        schema:
          type: integer
          format: int64
      - in: path
        name: invitacionId
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelInvitacionRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitacionResponse"
          description: Invitacion cancelada
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Falta el motivo
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: "Sin sesion, sin contexto activo o sin colaborador:manage"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: La invitacion no existe o es de otro tenant
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: La invitacion ya fue resuelta (invitacion-ya-resuelta)
      summary: Retirar una invitacion pendiente
      tags:
      - Invitaciones
  /api/v1/organizations/{orgId}/colaborador-invitaciones/{invitacionId}/resend:
    post:
      description: |-
        Emite un token nuevo y vuelve a mandar el correo. EL TOKEN ANTERIOR DEJA DE SERVIR en el mismo acto: dos enlaces validos para la misma invitacion dejan al invitado eligiendo cual usar.

        No crea una invitacion nueva. Conserva el id, la autoria y la fecha original, que es lo que hace que el listado siga diciendo desde cuando se espera respuesta.

        La respuesta NO trae el token: va al correo y a ningun lado mas.
      operationId: resendColaboradorInvitacion
      parameters:
      - in: path
        name: orgId
        required: true
        schema:
          type: integer
          format: int64
      - in: path
        name: invitacionId
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitacionResponse"
          description: Token rotado y correo encolado
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: "Sin sesion, sin contexto activo o sin colaborador:manage"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: La invitacion no existe o es de otro tenant. Los dos responden
            igual
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: "La invitacion ya fue aceptada, rechazada o cancelada (invitacion-ya-resuelta)"
      summary: Reenviar una invitacion pendiente
      tags:
      - Invitaciones
...
components:
  schemas:
...
    CancelInvitacionRequest:
      type: object
      description: Motivo con el que se retira una invitacion pendiente
      properties:
        reason:
          type: string
          description: Motivo declarado de la cancelacion
          example: Se equivoco de direccion de correo
          maxLength: 280
          minLength: 0
      required:
      - reason
...
    CreateInvitacionRequest:
      type: object
      description: Datos de la invitacion a emitir
      properties:
        consultorioId:
          type: integer
          format: int64
          description: Sede del vinculo propuesto. Ausente o null para alcance organizacion
          example: 3
        email:
          type: string
          format: email
          description: Email de la persona a invitar. Puede no tener cuenta todavia
          example: kine@centro.test
          maxLength: 320
          minLength: 0
        roleCode:
          type: string
          description: Rol con el que quedaria vinculada al aceptar
          example: PROFESIONAL
          maxLength: 64
          minLength: 0
      required:
      - email
      - roleCode
...
    CreateOrganizationRequest:
      type: object
      description: Datos para dar de alta una organizacion y su primer consultorio
      properties:
        name:
          type: string
          description: Nombre del centro
          example: Centro Kinesico Belgrano
          maxLength: 160
          minLength: 0
        planCode:
          type: string
          description: "Codigo del plan a contratar, del catalogo de GET /api/v1/plans"
          example: BASICO
          maxLength: 32
          minLength: 0
        slug:
          type: string
          description: "Identificador legible. Si se omite, se deriva del nombre y\
            \ se resuelve cualquier colision automaticamente"
          example: centro-kinesico-belgrano
          maxLength: 64
          minLength: 0
        timezone:
          type: string
          description: "Zona horaria IANA del centro. Si se omite, se usa la zona\
            \ por defecto de la plataforma"
          example: America/Argentina/Buenos_Aires
          maxLength: 64
          minLength: 0
      required:
      - name
      - planCode
...
    InvitacionResponse:
      type: object
      description: Invitacion emitida por la organizacion
      properties:
        aceptadaPorAccountId:
          type: integer
          format: int64
          description: "Cuenta que acepto, si alguien acepto"
        consultorioId:
          type: integer
          format: int64
          description: "Sede del vinculo propuesto, o null si es de toda la organizacion"
        createdAt:
          type: string
          format: date-time
        email:
          type: string
          example: kine@centro.test
        estado:
          type: string
          description: "PENDIENTE, ACEPTADA, RECHAZADA o CANCELADA"
          example: PENDIENTE
        expiraEn:
          type: string
          format: date-time
        id:
          type: integer
          format: int64
          example: 12
        invitadaPorAccountId:
          type: integer
          format: int64
          description: Cuenta del administrador que la emitio
        membershipId:
          type: integer
          format: int64
          description: "Vinculo creado al aceptar, si se acepto"
        organizationId:
          type: integer
          format: int64
          example: 1
        resolucionNota:
          type: string
          description: Motivo del rechazo o de la cancelacion
        resueltaEn:
          type: string
          format: date-time
          description: Cuando dejo de estar pendiente. Null mientras lo este
        roleCode:
          type: string
          example: PROFESIONAL
        vencida:
          type: boolean
          description: Si el enlace ya vencio. Puede ser true con estado PENDIENTE
        version:
          type: integer
          format: int64
          description: Version para el control de concurrencia optimista
...
    OrganizationResponse:
      type: object
      description: Organizacion (tenant) con su estado operativo efectivo
      properties:
        active:
          type: boolean
          description: false cuando la organizacion fue dada de baja logica
          example: true
        id:
          type: integer
          format: int64
          description: Identificador de la organizacion
          example: 1
        name:
          type: string
          description: Nombre del centro
          example: Centro Kinesico Belgrano
        operationalStatus:
          type: string
          description: "Estado operativo efectivo del tenant, derivado de la suscripcion\
            \ y de la baja logica de la organizacion"
          enum:
          - ACTIVA
          - SUSPENDIDA
          - CANCELADA
          - BAJA
          example: ACTIVA
        slug:
          type: string
          description: "Identificador legible y estable. No cambia al renombrar el\
            \ centro: se usa en URLs y en soporte"
          example: centro-kinesico-belgrano
        timezone:
          type: string
          description: Zona horaria IANA del centro
          example: America/Argentina/Buenos_Aires
        version:
          type: integer
          format: int64
          description: Version para bloqueo optimista. Devolverla tal cual en el PATCH
          example: 0
...
    ProblemDetail:
      type: object
      properties:
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties: {}
        status:
          type: integer
          format: int32
        title:
          type: string
        type:
          type: string
          format: uri
...